refactor: Amend naming to be more accurate
