+++ app/src/test/resources/bodyPoPBase64.txt
@@ -1,0 +1,1 @@
+eyJhbGciOiJFUzI1NiJ9.eyJpc3MiOiJpc3MiLCJhdWQiOiJhdWQiLCJleHAiOjAsImp0aSI6Imp0aSJ9
